Command line improvements
