+++ tools/cli/CTestTestfile.cmake
@@ -1,0 +1,8 @@
+# CMake generated Testfile for 
+# Source directory: /home/flackjacket/Desktop/Working Folder/ada_nn/ada/tools/cli
+# Build directory: /home/flackjacket/Desktop/Working Folder/ada_nn/ada/tools/cli
+# 
+# This file includes the relevant testing commands required for 
+# testing this directory and lists subdirectories to be tested as well.
+subdirs("../../_deps/cxxopts")
+subdirs("../../_deps/fmt")
