--- tools/output.txt
@@ -1,1 +1,0 @@
-By default, outputting to a file is disabled. Please enable -o flag (optionally with a path to activate it).
